clk: qcom: sdm845: Fix some clk stuck off during probe
When clocks are registered with the framework during the
probe phase, the clk_core_hold_state() function calls
clk_core_prepare_enable() for clocks that are already
enabled in hardware by the bootloader. Some clocks are
failing to activate at this point, leading to unnecessary
warnings in the kernel log.

The issue occurs specifically with certain GPU clocks,
which fail to turn on due to the required GDSC not being
enabled yet. However, this is not a functional problem,
as the GPU is not in use at this stage. Since these clocks
are not shared across multiple client devices, there's no
need for the framework to issue a proxy vote during
registration. To resolve this, the CLK_DONT_HOLD_STATE flag
is set for these clocks, preventing clk_core_prepare_enable()
from being called during registration.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>