msm: sde: Fix intf_sel for SDM845
In 4.14+ kernel the definition of the active interface for
cont splash was changed[1], which is invalid for older
SoCs (SDM845/660), for them a different bit[2] must be used
to check the active interface. Hence add a proper check for
these legacy SoCs to the get_ctl_intf function and protect
it by checking the SDE HW revision.

[1] https://git.codelinaro.org/clo/la/kernel/msm-4.14/-/commit/75cc47bfa5f8ffb8ee43d6597ac2e8c74093cd91
[2] https://git.codelinaro.org/clo/la/kernel/msm-4.9/-/blame/061c320644727afd4c9cfe5291dc3e7a4470e6cd/drivers/gpu/drm/msm/sde/sde_rm.c?page=2#L1433

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>