dts: qcom: sdm845-wsa881x: Update WSA enumeration
WCD SoundWire controller driver now requires explicit
definition of the number of secondary devices (codecs).

The WSA881x driver, requires explicitly specifying a prefix
name that will be used as the component name.

Additionally, by default, use the WSA8815 codec and disable
the WSA8810 codec because the connection with them is
established as with a codec, not as an aux device, and we
can no longer do this dynamically in the sound card driver.
The driver still supports both codec options, and if any
platform uses one or the other codec, this must be
explicitly specified in the corresponding overlay.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>